# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CredentialCreateRequest": {
        "properties": {
          "anti_abuse_pass": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anti Abuse Pass"
          },
          "consent_version": {
            "title": "Consent Version",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "purpose": {
            "enum": [
              "evaluation",
              "study_cabin",
              "qun_checkin"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "secret": {
            "format": "password",
            "title": "Secret",
            "type": "string",
            "writeOnly": true
          },
          "ttl_seconds": {
            "anyOf": [
              {
                "maximum": 2592000.0,
                "minimum": 300.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ttl Seconds"
          },
          "turnstile_token": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Token"
          },
          "upstream": {
            "enum": [
              "academic",
              "infospace",
              "qun100"
            ],
            "title": "Upstream",
            "type": "string"
          }
        },
        "required": [
          "upstream",
          "purpose",
          "secret",
          "consent_version"
        ],
        "title": "CredentialCreateRequest",
        "type": "object"
      },
...
      "CredentialResponse": {
        "properties": {
          "consent_version": {
            "title": "Consent Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Metadata",
            "type": "object"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CredentialStatus"
          },
          "upstream": {
            "title": "Upstream",
            "type": "string"
          }
        },
        "required": [
          "id",
          "upstream",
          "purpose",
          "status",
          "expires_at",
          "created_at",
          "last_used_at",
          "metadata",
          "consent_version"
        ],
        "title": "CredentialResponse",
        "type": "object"
      },
      "CredentialStatus": {
        "enum": [
          "active",
          "revoked",
          "deleted"
        ],
        "title": "CredentialStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Nanyee API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/credentials": {
      "get": {
        "operationId": "list_credentials",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CredentialResponse"
                  },
                  "title": "Response List Credentials",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Credentials",
        "tags": [
          "credentials"
        ]
      },
      "post": {
        "operationId": "create_credential",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Credential",
        "tags": [
          "credentials"
        ]
      }
    },
    "/api/v1/credentials/{credential_id}": {
      "delete": {
        "operationId": "revoke_credential",
        "parameters": [
          {
            "in": "path",
            "name": "credential_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Credential Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CredentialResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Credential",
        "tags": [
          "credentials"
        ]
      }
    },
...
  }
}